+++ Presentation.Web/Presentation.Web.csproj
@@ -384,14 +384,14 @@
     <Compile Include="Controllers\API\V2\External\ItSystems\Mapping\SystemDeletionConflictMappingExtensions.cs" />
     <Compile Include="Controllers\API\V2\Internal\DataProcessingRegistrations\DataProcessingRegistrationInternalV2Controller.cs" />
     <Compile Include="Controllers\API\V2\Internal\ItContracts\ItContractInternalV2Controller.cs" />
-    <Compile Include="Controllers\API\V2\Internal\ItSystems\ItSystemLocalOptionTypesInternalV2Controller.cs" />
+    <Compile Include="Controllers\API\V2\Internal\ItSystems\ItSystemLocalRegularOptionTypesInternalV2Controller.cs" />
     <Compile Include="Controllers\API\V2\Internal\ItSystems\ItSystemInternalV2Controller.cs" />
     <Compile Include="Controllers\API\V2\Internal\ItSystemUsages\ItSystemUsageInternalV2Controller.cs" />
     <Compile Include="Controllers\API\V2\Internal\ItSystemUsages\ItSystemUsageMigrationV2Controller.cs" />
     <Compile Include="Controllers\API\V2\Internal\ItSystemUsages\Mapping\IItSystemUsageMigrationResponseMapper.cs" />
     <Compile Include="Controllers\API\V2\Internal\ItSystemUsages\Mapping\ItSystemUsageMigrationResponseMapper.cs" />
-    <Compile Include="Controllers\API\V2\Internal\LocalRegularOptionTypesInternalV2Controller.cs" />
-    <Compile Include="Controllers\API\V2\Internal\LocalRoleOptionTypesInternalV2Controller.cs" />
+    <Compile Include="Controllers\API\V2\Internal\LocalOptionTypes\BaseLocalRegularOptionTypesInternalV2Controller.cs" />
+    <Compile Include="Controllers\API\V2\Internal\LocalOptionTypes\BaseLocalRoleOptionTypesInternalV2Controller.cs" />
     <Compile Include="Controllers\API\V2\Internal\Mapping\InternalDtoModelV2MappingExtensions.cs" />
     <Compile Include="Controllers\API\V2\Common\Helpers\ItSystemUsageV2QueryHandler.cs" />
     <Compile Include="Controllers\API\V2\Internal\Organizations\OrganizationGridInternalV2Controller.cs" />
@@ -401,7 +401,7 @@
     <Compile Include="Controllers\API\V2\Internal\OrganizationUnits\Mapping\OrganizationUnitOriginMappingExtensions.cs" />
     <Compile Include="Controllers\API\V2\Internal\OrganizationUnits\Mapping\OrganizationUnitResponseModelMapper.cs" />
     <Compile Include="Controllers\API\V2\Internal\OrganizationUnits\Mapping\OrganizationUnitWriteModelMapper.cs" />
-    <Compile Include="Controllers\API\V2\Internal\OrganizationUnits\OrganizationUnitLocalOptionTypesInternalV2Controller.cs" />
+    <Compile Include="Controllers\API\V2\Internal\OrganizationUnits\OrganizationUnitLocalRoleOptionTypesInternalV2Controller.cs" />
     <Compile Include="Controllers\API\V2\Internal\OrganizationUnits\OrganizationUnitRegistrationInternalV2Controller.cs" />
     <Compile Include="Controllers\API\V2\Internal\OrganizationUnits\OrganizationUnitsInternalV2Controller.cs" />
     <Compile Include="Controllers\API\V2\Internal\Sts\StsOrganizationSynchronizationInternalV2Controller.cs" />
